feat: kafka producer ack 설정

+++ init.sql
@@ -2,20 +2,20 @@ CREATE DATABASE IF NOT EXISTS stockly;
 USE stockly;
 
 CREATE TABLE IF NOT EXISTS stock (
-id INT AUTO_INCREMENT PRIMARY KEY,
-created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-is_deleted TINYINT DEFAULT 0,
-symbol VARCHAR(100),
-high INT,
-low INT,
-volume BIGINT,
-date DATETIME,
-open INT,
-close INT,
-rate DOUBLE,
-rate_price INT,
-trading_value BIGINT
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
+    is_deleted TINYINT DEFAULT 0,
+    symbol VARCHAR(100),
+    high INT,
+    low INT,
+    volume BIGINT,
+    date DATETIME,
+    open INT,
+    close INT,
+    rate DOUBLE,
+    rate_price INT,
+    trading_value DECIMAL(20, 2)
 );
 
 ALTER TABLE stock
